DaysEarned is decimal

+++ packages/api/AlvTime.Business/Absence/CustomVacationOverrideOverview.cs
@@ -4,5 +4,5 @@ public class CustomVacationOverrideOverview
 {
     public int UserId { get; set; }
     public int Year { get; set; }
-    public int DaysEarned { get; set; }
+    public decimal DaysEarned { get; set; }
 }
