+++ packages/api/AlvTime.Persistence/DatabaseModels/VacationDaysEarnedOverride.cs
@@ -7,7 +7,7 @@ namespace AlvTime.Persistence.DatabaseModels
     {
         public int Id { get; set; }
         public int UserId { get; set; }
-        public int DaysEarned { get; set; }
+        public decimal DaysEarned { get; set; }
         public int Year { get; set; }
 
         public virtual User User { get; set; }
